OCIOv2: Fix compilations issues in Cycles.
This is a rough version. The result should be slower than we have. We
should find a solution where we could cache the device_processor or find
a better way how to increase the performance by evaluating an area or
using GPU processors. OCIO also has a function called
`getOptimizedCPUProcessor`.