Tracking: Implement proportional editing
Disclaimer: the user interface and interaction is not final. There
are known sharp corners, which are planned to be addressed while
the patch is being tested by people.

The idea is to avoid sudden jumps of track path when artists detects
track got slid away from the feature and one re-positions track back.
The implementation is based on a proportional editing, which allows
to control whether to transform single marker or whether to propagate
transform to markers from neighbour frames. The proportional editing
also allows to control propagation curve. The propagation will stop
at the end of tracked segment.

To test the new functionality:

1. Enable proportional editing
2. Enable "Connected Only". This is essential, as otherwise the new
   functionality is not enabled.
3. Grab marker.

Open topics:
- Make the functionality more discoverable.
- Make visualization of how far propagation goes more clear.
- Make defaults more usable.

Technical TODO:
- Most likely CANCEL of the transform does not properly restore the
  full state prior to the transform. This needs to be solved.

Differential Revision: https://developer.blender.org/D2771